refactor(control-ui): split agent menu, identity actions, model staging out of oversized files
