+++ ui/src/app/bootstrap.ts
@@ -36,11 +36,11 @@ import {
   loadSettings,
   patchSettings,
   persistSessionToken,
-  resolveApplicationStartupSettings,
   resolvePageGatewaySettings,
   saveSettings,
   type UiSettings,
 } from "./settings.ts";
+import { resolveApplicationStartupSettings } from "./startup-settings.ts";
 import { startThemeTransition } from "./theme-transition.ts";
 import { resolveTheme, type ThemeMode } from "./theme.ts";
 import { createWebPushCapability } from "./web-push.ts";
