+++ ui/src/app/settings.node.test.ts
@@ -8,10 +8,10 @@ import {
   persistSessionToken,
   resetUnpersistedSettingsForTest,
   resolvePageGatewaySettings,
-  resolveApplicationStartupSettings,
   saveSettings,
   type UiSettings,
 } from "./settings.ts";
+import { resolveApplicationStartupSettings } from "./startup-settings.ts";
 
 function setTestLocation(params: { protocol: string; host: string; pathname: string }) {
   vi.stubGlobal("location", {
