+++ ui/src/components/app-sidebar-agent-menu.ts
@@ -1,0 +1,258 @@
+// Sidebar agent-chip menu (switcher, filter, utility rows), split out of
+// app-sidebar.ts to keep that hot component inside the TS LOC ratchet.
+import { html, nothing } from "lit";
+import { titleForRoute, type NavigationRouteId } from "../app-navigation.ts";
+import type { ApplicationNavigationOptions } from "../app/context.ts";
+import type { ThemeMode } from "../app/theme.ts";
+import { t } from "../i18n/index.ts";
+import { normalizeAgentLabel, resolveAgentTextAvatar } from "../lib/agents/display.ts";
+import { buildExternalLinkRel, EXTERNAL_LINK_TARGET } from "../lib/external-link.ts";
+import { normalizeAgentId } from "../lib/sessions/session-key.ts";
+import { icons, type IconName } from "./icons.ts";
+
+// External rows of the footer agent menu. Docs-first: public docs pages over
+// raw GitHub, matching the ClawSweeper docs-link policy for user-facing copy.
+const AGENT_MENU_LINKS: ReadonlyArray<{ href: string; icon: IconName; label: () => string }> = [
+  { href: "https://docs.openclaw.ai", icon: "book", label: () => t("common.docs") },
+  {
+    href: "https://docs.openclaw.ai/help",
+    icon: "messageSquare",
+    label: () => t("agentChip.getHelp"),
+  },
+  { href: "https://discord.gg/clawd", icon: "users", label: () => t("agentChip.discord") },
+  {
+    href: "https://docs.openclaw.ai/releases",
+    icon: "scrollText",
+    label: () => t("agentChip.viewChangelog"),
+  },
+];
+
+/** Above this roster size the chip menu switches to pinned agents + filter. */
+export const QUICK_SWITCH_AGENT_LIMIT = 10;
+
+type AgentMenuAgent = { id: string; name?: string; identity?: { name?: string; emoji?: string } };
+
+export type SidebarAgentMenuParams = {
+  position: { x: number; bottom: number } | null;
+  activeId: string;
+  activeName: string;
+  agents: readonly AgentMenuAgent[];
+  filter: string;
+  pinnedAgentIds: readonly string[];
+  connected: boolean;
+  canPairDevice: boolean;
+  basePath: string;
+  gatewayVersion: string | null;
+  themeMode: ThemeMode;
+  agentUnreadCount: (agentId: string) => number;
+  onFilterChange: (next: string) => void;
+  onSwitchAgent: (agentId: string) => void;
+  onAskCapabilities: (agentId: string) => void;
+  onClose: () => void;
+  onNavigate: (routeId: NavigationRouteId, options?: ApplicationNavigationOptions) => void;
+  onPairMobile: () => void;
+};
+
+/** Rows for the chip switcher. Small rosters list everything; past
+    QUICK_SWITCH_AGENT_LIMIT the menu shows pinned agents (plus the active
+    one) and the filter searches the full roster. */
+export function sidebarAgentMenuRows(params: {
+  agents: readonly AgentMenuAgent[];
+  activeId: string;
+  filter: string;
+  pinnedAgentIds: readonly string[];
+}) {
+  const { agents, activeId } = params;
+  const pinnedIds = new Set(params.pinnedAgentIds.map((agentId) => normalizeAgentId(agentId)));
+  const sorted = agents.toSorted((a, b) => {
+    const aPinned = pinnedIds.has(normalizeAgentId(a.id)) ? 0 : 1;
+    const bPinned = pinnedIds.has(normalizeAgentId(b.id)) ? 0 : 1;
+    return aPinned - bPinned;
+  });
+  if (agents.length <= QUICK_SWITCH_AGENT_LIMIT) {
+    return { rows: sorted, showFilter: false };
+  }
+  const query = params.filter.trim().toLowerCase();
+  if (query) {
+    const rows = sorted.filter((entry) => {
+      const agentId = normalizeAgentId(entry.id);
+      return (
+        agentId.toLowerCase().includes(query) ||
+        normalizeAgentLabel(entry).toLowerCase().includes(query)
+      );
+    });
+    return { rows, showFilter: true };
+  }
+  const rows =
+    pinnedIds.size > 0
+      ? sorted.filter((entry) => {
+          const agentId = normalizeAgentId(entry.id);
+          return pinnedIds.has(agentId) || agentId === activeId;
+        })
+      : sorted.slice(0, QUICK_SWITCH_AGENT_LIMIT);
+  return { rows, showFilter: true };
+}
+
+function renderAgentRow(agent: AgentMenuAgent, params: SidebarAgentMenuParams) {
+  const agentId = normalizeAgentId(agent.id);
+  const label = normalizeAgentLabel(agent);
+  const active = agentId === params.activeId;
+  const unread = active ? 0 : params.agentUnreadCount(agentId);
+  const initial = resolveAgentTextAvatar(agent) ?? (label || agent.id).slice(0, 1).toUpperCase();
+  return html`
+    <button
+      type="button"
+      class="sidebar-customize-menu__item"
+      role="menuitemradio"
+      tabindex="-1"
+      aria-checked=${String(active)}
+      @click=${() => params.onSwitchAgent(agentId)}
+    >
+      <span class="sidebar-agent-section__avatar" aria-hidden="true">${initial}</span>
+      <span class="sidebar-customize-menu__text">${label}</span>
+      ${unread > 0
+        ? html`<span
+            class="session-unread-dot"
+            role="img"
+            aria-label=${t("sessionsView.unread")}
+          ></span>`
+        : nothing}
+      <span class="sidebar-customize-menu__check" aria-hidden="true">
+        ${active ? icons.check : nothing}
+      </span>
+    </button>
+  `;
+}
+
+export function renderSidebarAgentMenu(params: SidebarAgentMenuParams) {
+  const position = params.position;
+  if (!position) {
+    return nothing;
+  }
+  const { activeId, activeName, agents } = params;
+  const { rows, showFilter } = sidebarAgentMenuRows(params);
+  return html`
+    <openclaw-menu-surface>
+      <div
+        class="sidebar-customize-menu sidebar-agent-menu"
+        role="menu"
+        aria-label=${t("agentChip.menuLabel")}
+        style="left: ${position.x}px; bottom: ${position.bottom}px;"
+      >
+        ${agents.length > 1
+          ? html`
+              <div class="sidebar-customize-menu__title">${t("agentChip.agents")}</div>
+              ${showFilter
+                ? html`
+                    <div class="sidebar-agent-menu__filter">
+                      <input
+                        type="text"
+                        .value=${params.filter}
+                        placeholder=${t("agentChip.filterAgents")}
+                        aria-label=${t("agentChip.filterAgents")}
+                        @input=${(event: Event) =>
+                          params.onFilterChange((event.target as HTMLInputElement).value)}
+                      />
+                    </div>
+                  `
+                : nothing}
+              <div class="sidebar-agent-menu__list">
+                ${rows.map((entry) => renderAgentRow(entry, params))}
+                ${rows.length === 0
+                  ? html`<div class="sidebar-agent-menu__empty">
+                      ${t("agentChip.noAgentMatches")}
+                    </div>`
+                  : nothing}
+              </div>
+              <div class="sidebar-customize-menu__separator" role="separator"></div>
+            `
+          : nothing}
+        <button
+          type="button"
+          class="sidebar-customize-menu__item"
+          role="menuitem"
+          tabindex="-1"
+          ?disabled=${!params.connected}
+          @click=${() => params.onAskCapabilities(activeId)}
+        >
+          <span class="nav-item__icon" aria-hidden="true">${icons.bot}</span>
+          <span class="sidebar-customize-menu__text">
+            ${t("agentChip.whatCanAgentDo", { name: activeName })}
+          </span>
+        </button>
+        <button
+          type="button"
+          class="sidebar-customize-menu__item"
+          role="menuitem"
+          tabindex="-1"
+          @click=${() => {
+            params.onClose();
+            params.onNavigate("agents", { search: `?agent=${encodeURIComponent(activeId)}` });
+          }}
+        >
+          <span class="nav-item__icon" aria-hidden="true">${icons.users}</span>
+          <span class="sidebar-customize-menu__text">${t("agentChip.agentSettings")}</span>
+        </button>
+        <div class="sidebar-customize-menu__separator" role="separator"></div>
+        <button
+          type="button"
+          class="sidebar-customize-menu__item"
+          role="menuitem"
+          tabindex="-1"
+          @click=${() => {
+            params.onClose();
+            params.onNavigate("config");
+          }}
+        >
+          <span class="nav-item__icon" aria-hidden="true">${icons.settings}</span>
+          <span class="sidebar-customize-menu__text">${titleForRoute("config")}</span>
+        </button>
+        <button
+          type="button"
+          class="sidebar-customize-menu__item sidebar-pair-mobile"
+          role="menuitem"
+          tabindex="-1"
+          ?disabled=${!params.canPairDevice}
+          title=${params.canPairDevice ? nothing : t("nodes.pairing.adminRequired")}
+          @click=${() => {
+            params.onClose();
+            params.onPairMobile();
+          }}
+        >
+          <span class="nav-item__icon" aria-hidden="true">${icons.smartphone}</span>
+          <span class="sidebar-customize-menu__text">${t("nodes.pairing.button")}</span>
+        </button>
+        ${AGENT_MENU_LINKS.map(
+          (link) => html`
+            <a
+              class="sidebar-customize-menu__item"
+              role="menuitem"
+              tabindex="-1"
+              href=${link.href}
+              target=${EXTERNAL_LINK_TARGET}
+              rel=${buildExternalLinkRel()}
+              @click=${() => params.onClose()}
+            >
+              <span class="nav-item__icon" aria-hidden="true">${icons[link.icon]}</span>
+              <span class="sidebar-customize-menu__text">${link.label()}</span>
+            </a>
+          `,
+        )}
+        <div class="sidebar-customize-menu__separator" role="separator"></div>
+        <div class="sidebar-agent-menu__footer">
+          <openclaw-sidebar-build-chip
+            .basePath=${params.basePath}
+            .gatewayVersion=${params.gatewayVersion}
+            .onNavigate=${(routeId: "about") => {
+              params.onClose();
+              params.onNavigate(routeId);
+            }}
+          ></openclaw-sidebar-build-chip>
+          <span class="sidebar-mode-switch">
+            <openclaw-theme-mode-toggle .mode=${params.themeMode}></openclaw-theme-mode-toggle>
+          </span>
+        </div>
+      </div>
+    </openclaw-menu-surface>
+  `;
+}
